Refactor redis streams data structure to sep crate
