+++ micro-judge/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "micro_judge"
-version = "0.1.4"
+version = "0.1.5"
 authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
 
@@ -9,3 +9,4 @@ uuid = { version = "0.8.1", features = ["serde", "v4"] }
 bincode = "1.2.1"
 micro_model_moves = { path = "../micro-model/moves" }
 redis_conn_pool = { path = "../redis-conn-pool" }
+redis_streams = {path = "../redis_streams"}
