+++ micro-judge/src/lib.rs
@@ -1,8 +1,7 @@
 extern crate micro_model_moves;
-
 pub extern crate redis_conn_pool;
+extern crate redis_streams;
 extern crate uuid;
-
 
 pub mod game;
 pub mod io;
