+++ redis_streams/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "redis_streams"
+version = "0.1.0"
+authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+uuid = {version="0.8.1",features = ["serde","v4"]}
